Move grid to its own component.
